+++ src/modules/style/winapi_color.rs
@@ -101,6 +101,10 @@ impl ITerminalColor for WinApiColor {
                     Color::DarkCyan => fg_green | fg_blue,
                     Color::Grey => fg_intensity,
                     Color::White => fg_intensity | fg_red | fg_green | fg_blue,
+
+                    /* WinApi will be used for systems that do not support ANSI, those are windows version less then 10. RGB and 255 (AnsiBValue) colors are not supported in that case.*/
+                    Color::Rgb{ r, g, b } => { 0 }
+                    Color::AnsiValue(val) => { 0 }
                 };
             }
             ColorType::Background => {
@@ -120,6 +124,10 @@ impl ITerminalColor for WinApiColor {
                     Color::DarkCyan => bg_green | bg_blue,
                     Color::Grey => bg_intensity,
                     Color::White => bg_intensity | bg_red | bg_green | bg_blue,
+
+                    /* WinApi will be used for systems that do not support ANSI, those are windows version less then 10. RGB and 255 (AnsiBValue) colors are not supported in that case.*/
+                    Color::Rgb{ r, g, b } => { 0 }
+                    Color::AnsiValue(val) => { 0 }
                 };
             }
         };